Revert "Process Manager Refactoring (#297)" (#298)
This reverts commit 6465faebe268bffbba85ef73da8f7a6ef3bf996b.
