+++ src/tracer/src/extracts/ebpf_watcher/watcher.rs
@@ -39,8 +39,11 @@ impl EbpfWatcher {
     }
 
     pub async fn update_targets(self: &Arc<Self>, targets: Vec<Target>) -> Result<()> {
-        let process_manager = self.process_manager.write().await;
-        process_manager.update_targets(targets).await?;
+        self.process_manager
+            .write()
+            .await
+            .update_targets(targets)
+            .await?;
         Ok(())
     }
 
@@ -159,7 +162,7 @@ impl EbpfWatcher {
         match tokio::runtime::Handle::try_current() {
             Ok(_) => {
                 tokio::spawn(async move {
-                    let process_manager = self.process_manager.write().await;
+                    let mut process_manager = self.process_manager.write().await;
                     process_manager.set_ebpf_task(task).await;
                 });
                 info!("eBPF monitoring task initialized successfully");
