+++ src/tracer/src/extracts/process/mod.rs
@@ -1,7 +1,4 @@
-pub mod extract_process_data;
+mod extract_process_data;
 pub mod process_manager;
 pub mod process_utils;
-pub mod types;
-
-// Re-export the main ProcessManager for convenience
-pub use process_manager::ProcessManager;
+mod types;
