--- src/tracer/src/extracts/process/process_manager/handlers/mod.rs
@@ -1,3 +1,0 @@
-pub mod oom;
-pub mod process_starts;
-pub mod process_terminations;
