--- src/tracer/src/extracts/process/process_manager/mod.rs
@@ -1,10 +1,0 @@
-pub mod handlers;
-pub mod logger;
-pub mod manager;
-pub mod matcher;
-pub mod metrics;
-pub mod state;
-pub mod system_refresher;
-
-// Re-export the main ProcessManager for convenience
-pub use manager::ProcessManager;
